flask-crud-app/backend/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
pytest
